# backend/src/main/resources/openapi/sims-openapi.yaml
openapi: 3.0.3
info:
  title: SIMS API — Exam, Fee and Communication modules
  description: |
    School Information Management System — the exam, fee and communication modules
    after the business-workflow redesign.

    ## Conventions

    **Nested business objects.** Responses embed the data a screen needs rather than
    bare foreign keys: `subject`, `student`, `schoolClass`, `createdBy` and
    `updatedBy` are objects carrying ids *and* names. The old flat `*Id` fields are
    still returned but are deprecated and will be removed once the React app has
    migrated off them.

    **Pagination.** List endpoints return a Spring `Page`:
    `?page=0&size=20&sort=createdAt,desc`.

    **Actor fields.** Requests carry `actorId` (admin staff) or `enteredById` /
    `updatedById` (teacher) because authentication is still being built by another
    developer. These fields disappear once the JWT principal is available.

    **Marks and fee "sheets".** Marks and fee statuses are worked on a section at a
    time, so each has a `GET /sheet` that returns the whole roster pre-filled and a
    `PUT /sheet` that saves it in one transaction.
  version: 2.0.0
  contact:
    name: SIMS Team
servers:
  - url: http://localhost:8081
    description: Local development
  - url: https://api.example.com
    description: Production
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/marks/sheet:
    get:
      tags: [Marks]
      summary: Marks-entry sheet
      description: |
        The whole marks-entry screen in one call: exam, subject, section and the full
        student roster (ordered by roll number) with any existing marks pre-filled.
        `marksObtained` is `null` for students not yet marked.

        Replaces roughly eight lookups the UI previously had to make.
      operationId: getMarksSheet
      parameters:
        - name: examSubjectId
          in: query
          required: true
          schema: { type: string, format: uuid }
        - name: sectionId
          in: query
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Marks sheet
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MarksSheetResponse' }
        '400':
          description: The section does not belong to the exam subject's class
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }

    put:
      tags: [Marks]
      summary: Save a marks sheet
      description: |
        Transactional bulk upsert. The entire payload is validated before anything is
        written, so one bad row cannot leave the sheet half-saved. A `null`
        `marksObtained` clears a previously recorded mark.

        Rejected when the exam is published.
      operationId: saveMarksSheet
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MarksSheetSaveRequest' }
      responses:
        '200':
          description: Sheet saved; the refreshed sheet is returned
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MarksSheetResponse' }
        '400':
          description: |
            Validation failure, a mark above `maxMarks`, a duplicate enrollment in the
            payload, or an enrollment outside the section.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: The exam is published, so marks are frozen
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /api/v1/marks/results/student:
    get:
      tags: [Marks]
      summary: One student's result
      description: |
        Every subject of the exam for the student's class, with marks, per-subject
        percentage, totals and overall percentage. Unmarked subjects appear with a
        `null` mark so the report card shows the gap.
      operationId: getStudentResult
      parameters:
        - name: examId
          in: query
          required: true
          schema: { type: string, format: uuid }
        - name: enrollmentId
          in: query
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Student result
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StudentResultResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /api/v1/marks/results/section:
    get:
      tags: [Marks]
      summary: Result summary for a section
      description: One result block per student in the section, bounded by roster size.
      operationId: getSectionResults
      parameters:
        - name: examId
          in: query
          required: true
          schema: { type: string, format: uuid }
        - name: sectionId
          in: query
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Section results
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/StudentResultResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }

  # ==================================================================
  # Marks — CRUD
  # ==================================================================
  /api/v1/marks:
    get:
      tags: [Marks]
      summary: List marks
      operationId: getAllMarks
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200':
          description: Page of marks
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PageMarkResponse' }
        '500': { $ref: '#/components/responses/InternalServerError' }

    post:
      tags: [Marks]
      summary: Record a single mark
      description: |
        Single-record entry, kept for corrections. Bulk entry should use
        `PUT /api/v1/marks/sheet`.
      operationId: createMark
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MarkRequest' }
      responses:
        '200':
          description: Mark recorded
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MarkResponse' }
        '400':
          description: |
            Validation failure, a mark above `maxMarks`, or a student not enrolled in
            the exam subject's class.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Mark already recorded for this pair, or the exam is published
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /api/v1/marks/{id}:
    parameters:
      - $ref: '#/components/parameters/PathId'
    get:
      tags: [Marks]
      summary: Get mark by id
      operationId: getMarkById
      responses:
        '200':
          description: Mark
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MarkResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }

    put:
      tags: [Marks]
      summary: Correct a mark
      operationId: updateMark
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MarkRequest' }
      responses:
        '200':
          description: Mark updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MarkResponse' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Duplicate pair, or the exam is published
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '500': { $ref: '#/components/responses/InternalServerError' }

    delete:
      tags: [Marks]
      summary: Delete a mark
      operationId: deleteMark
      responses:
        '204': { description: Mark deleted }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: The exam is published, so marks are frozen
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /api/v1/marks/exam-subject/{examSubjectId}:
    parameters:
      - name: examSubjectId
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Marks]
      summary: Marks for an exam subject
      description: Prefer `GET /api/v1/marks/sheet` for the entry screen.
      operationId: getMarksByExamSubject
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: Page of marks
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PageMarkResponse' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /api/v1/marks/enrollment/{enrollmentId}:
    parameters:
      - name: enrollmentId
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Marks]
      summary: Marks for a student enrollment
      description: Bounded by the student's subject count, so returned unpaginated.
      operationId: getMarksByEnrollment
      responses:
        '200':
          description: Marks
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/MarkResponse' }
        '500': { $ref: '#/components/responses/InternalServerError' }

  # ==================================================================
  # Fees
  # ==================================================================
  /api/v1/fees:
    get:
      tags: [Fees]
      summary: List fees
      operationId: getAllFees
      parameters:
        - $ref: '#/components/parameters/ClassIdFilter'
        - $ref: '#/components/parameters/AcademicYearIdFilter'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200':
          description: Page of fees
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PageFeeResponse' }
        '500': { $ref: '#/components/responses/InternalServerError' }

    post:
      tags: [Fees]
      summary: Define a fee
      description: Unique per (class, academic year, term).
      operationId: createFee
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FeeRequest' }
      responses:
        '200':
          description: Fee created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FeeResponse' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: A fee already exists for this class, year and term
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /api/v1/fees/{id}:
    parameters:
      - $ref: '#/components/parameters/PathId'
    get:
      tags: [Fees]
      summary: Get fee by id
      operationId: getFeeById
      responses:
        '200':
          description: Fee
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FeeResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }

    put:
      tags: [Fees]
      summary: Update a fee
      description: |
        Changing the class or academic year is rejected once student statuses exist,
        since it would move those students' records onto a different class or year.
      operationId: updateFee
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FeeRequest' }
      responses:
        '200':
          description: Fee updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FeeResponse' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Duplicate definition, or statuses already recorded
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '500': { $ref: '#/components/responses/InternalServerError' }

    delete:
      tags: [Fees]
      summary: Delete a fee
      description: Rejected while any student fee status references it.
      operationId: deleteFee
      responses:
        '204': { description: Fee deleted }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Student fee statuses are recorded against this fee
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '500': { $ref: '#/components/responses/InternalServerError' }

  # ==================================================================
  # Student Fee Status
  # ==================================================================
...
components:
  parameters:
    PathId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    Page:
      name: page
      in: query
      required: false
      schema: { type: integer, minimum: 0, default: 0 }
    Size:
      name: size
      in: query
      required: false
      schema: { type: integer, minimum: 1, default: 20 }
    Sort:
      name: sort
      in: query
      required: false
      description: '`property,(asc|desc)`; repeatable.'
      schema:
        type: array
        items: { type: string }
      example: ['createdAt,desc']
    ClassIdFilter:
      name: classId
      in: query
      required: false
      schema: { type: string, format: uuid }
    AcademicYearIdFilter:
      name: academicYearId
      in: query
      required: false
      schema: { type: string, format: uuid }
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT issued by the login endpoint. Authentication is still being built, so
        endpoints are currently open and role rules are recorded as TODOs beside each
        controller method.
  responses:
...
    NotFound:
      description: A referenced record does not exist
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            timestamp: '2026-08-02T10:30:00Z'
            status: 404
            error: Not Found
            message: 'Exam not found with id: 9f1c8b2e-...'
            path: /api/v1/exams/9f1c8b2e-...
    ValidationError:
      description: Bean validation failure; offending fields are listed in `fieldErrors`
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            timestamp: '2026-08-02T10:30:00Z'
            status: 400
            error: Bad Request
            message: Validation failed
            path: /api/v1/exams
            fieldErrors:
              examName: Exam name is required
              actorId: Acting admin id is required
    InternalServerError:
      description: Unexpected server error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
...
    ExamStatus:
      type: string
      enum: [DRAFT, PUBLISHED]
...
    SubjectType:
      type: string
      enum: [COMPULSORY, ELECTIVE]

    # ---------------- shared summaries ----------------
    ActorSummary:
      type: object
      description: A person who created or last changed a record.
      properties:
        id: { type: string, format: uuid }
        name: { type: string, example: R. Iyer }
        role:
          type: string
          description: Designation, e.g. PRINCIPAL or SENIOR_TEACHER.
          example: SENIOR_TEACHER
    ClassSummary:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string, example: Grade 10 }
    SubjectSummary:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string, example: Mathematics }
        code: { type: string, example: MATH001 }
        type: { $ref: '#/components/schemas/SubjectType' }
    AcademicYearSummary:
      type: object
      properties:
        id: { type: string, format: uuid }
        label: { type: string, example: 2025-2026 }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
    SectionSummary:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string, example: A }
        classId: { type: string, format: uuid }
        className: { type: string, example: Grade 10 }
        displayName:
          type: string
          description: Ready-made heading label.
          example: Grade 10 - A
    StudentSummary:
      type: object
      description: A student seen through their enrollment.
      properties:
        enrollmentId: { type: string, format: uuid }
        studentId: { type: string, format: uuid }
        name: { type: string, example: Aarav Sharma }
        rollNumber: { type: string, example: '04' }
        admissionNumber: { type: string, example: ADM-2025-0142 }

    # ---------------- errors & paging ----------------
    ErrorResponse:
      type: object
      properties:
        timestamp: { type: string, format: date-time }
        status: { type: integer, example: 409 }
        error: { type: string, example: Conflict }
        message: { type: string }
        path: { type: string }
        fieldErrors:
          type: object
          nullable: true
          additionalProperties: { type: string }
          description: Present only for bean validation failures.
      required: [timestamp, status, error, message, path]
    PageMetadata:
      type: object
      properties:
        totalElements: { type: integer, format: int64, example: 137 }
        totalPages: { type: integer, example: 7 }
        number: { type: integer, description: Zero-based page index, example: 0 }
        size: { type: integer, example: 20 }
        numberOfElements: { type: integer, example: 20 }
        first: { type: boolean }
        last: { type: boolean }
        empty: { type: boolean }

    # ---------------- exam ----------------
    ExamSummary:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string, example: Term 1 Examination }
        status: { $ref: '#/components/schemas/ExamStatus' }
...
    MarkRequest:
      type: object
      required: [enrollmentId, examSubjectId, marksObtained, enteredById]
      properties:
        enrollmentId: { type: string, format: uuid }
        examSubjectId: { type: string, format: uuid }
        marksObtained: { type: number, format: double, minimum: 0, example: 85.5 }
        enteredById:
          type: string
          format: uuid
          description: |
            Teacher performing the action — becomes `enteredBy` on create and
            `updatedBy` on update. Removed once authentication lands.
    MarkResponse:
      type: object
      properties:
        markId: { type: string, format: uuid }
        marksObtained: { type: number, format: double, example: 85.5 }
        maxMarks: { type: integer, example: 100 }
        student: { $ref: '#/components/schemas/StudentSummary' }
        subject: { $ref: '#/components/schemas/SubjectSummary' }
        exam: { $ref: '#/components/schemas/ExamSummary' }
        enteredBy: { $ref: '#/components/schemas/ActorSummary' }
        updatedBy: { $ref: '#/components/schemas/ActorSummary' }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        enrollmentId: { type: string, format: uuid, deprecated: true }
        examSubjectId: { type: string, format: uuid, deprecated: true }
        enteredById: { type: string, format: uuid, deprecated: true }
    MarksSheetEntry:
      type: object
      properties:
        student: { $ref: '#/components/schemas/StudentSummary' }
        markId:
          type: string
          format: uuid
          nullable: true
          description: Null until a mark has been saved.
        marksObtained:
          type: number
          format: double
          nullable: true
          description: Null when not yet entered.
        enteredBy: { $ref: '#/components/schemas/ActorSummary' }
        updatedAt: { type: string, format: date-time, nullable: true }
    MarksSheetResponse:
      type: object
      properties:
        examSubjectId: { type: string, format: uuid }
        exam: { $ref: '#/components/schemas/ExamSummary' }
        subject: { $ref: '#/components/schemas/SubjectSummary' }
        section: { $ref: '#/components/schemas/SectionSummary' }
        maxMarks: { type: integer, example: 100 }
        complete:
          type: boolean
          description: True once every student in the roster has a mark.
        totalStudents: { type: integer, example: 40 }
        marksEntered: { type: integer, example: 37 }
        editable:
          type: boolean
          description: False when the exam is published; render the grid read-only.
        entries:
          type: array
          items: { $ref: '#/components/schemas/MarksSheetEntry' }
    MarksSheetEntryRequest:
      type: object
      required: [enrollmentId]
      properties:
        enrollmentId: { type: string, format: uuid }
        marksObtained:
          type: number
          format: double
          nullable: true
          minimum: 0
          description: Null clears a previously entered mark.
    MarksSheetSaveRequest:
      type: object
      required: [examSubjectId, sectionId, enteredById, entries]
      properties:
        examSubjectId: { type: string, format: uuid }
        sectionId: { type: string, format: uuid }
        enteredById: { type: string, format: uuid }
        entries:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/MarksSheetEntryRequest' }
...
    StudentResultSubject:
      type: object
      properties:
        examSubjectId: { type: string, format: uuid }
        subject: { $ref: '#/components/schemas/SubjectSummary' }
        maxMarks: { type: integer }
        marksObtained:
          type: number
          format: double
          nullable: true
          description: Null when the subject has not been marked for this student.
        percentage: { type: number, format: double, nullable: true, example: 85.50 }
    StudentResultResponse:
      type: object
      properties:
        exam: { $ref: '#/components/schemas/ExamSummary' }
        student: { $ref: '#/components/schemas/StudentSummary' }
        section: { $ref: '#/components/schemas/SectionSummary' }
        subjects:
          type: array
          items: { $ref: '#/components/schemas/StudentResultSubject' }
        totalMarksObtained: { type: number, format: double, example: 462.5 }
        totalMaxMarks: { type: integer, example: 500 }
        percentage:
          type: number
          format: double
          nullable: true
          description: Across marked subjects only, 2 dp. Null when nothing is marked.
          example: 92.50
        subjectsMarked: { type: integer, example: 5 }
        subjectsTotal: { type: integer, example: 6 }

    # ---------------- fee ----------------
...
    FeeRequest:
      type: object
      required: [classId, academicYearId, termName, actorId]
      properties:
        classId: { type: string, format: uuid }
        academicYearId: { type: string, format: uuid }
        termName: { type: string, maxLength: 255, example: Term 1 }
        actorId: { type: string, format: uuid }
    FeeResponse:
      type: object
      properties:
        feeId: { type: string, format: uuid }
        termName: { type: string }
        schoolClass: { $ref: '#/components/schemas/ClassSummary' }
        academicYear: { $ref: '#/components/schemas/AcademicYearSummary' }
        updatedBy: { $ref: '#/components/schemas/ActorSummary' }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        classId: { type: string, format: uuid, deprecated: true }
        academicYearId: { type: string, format: uuid, deprecated: true }
...
    PageMarkResponse:
      allOf:
        - $ref: '#/components/schemas/PageMetadata'
        - type: object
          properties:
            content:
              type: array
              items: { $ref: '#/components/schemas/MarkResponse' }
    PageFeeResponse:
      allOf:
        - $ref: '#/components/schemas/PageMetadata'
        - type: object
          properties:
            content:
              type: array
              items: { $ref: '#/components/schemas/FeeResponse' }
...